--- src/reducer/utils.js
@@ -1,3 +1,0 @@
-export const extend = (prevState, newState) => {
-  return Object.assign({}, prevState, newState);
-};
